🚚 Move : 업체 서비스 폴더 구조 정리
Related to: #55
