--- company-service/src/main/java/com/team/companyservice/application/company/UpdateCompanyActiveRequest.java
@@ -1,11 +1,0 @@
-package com.team.companyservice.application.company;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Getter
-public class UpdateCompanyActiveRequest {
-
-    @NotNull
-    private Boolean isActive;
-}
