+++ company-service/src/main/java/com/team/companyservice/company/application/dto/request/UpdateCompanyActiveRequest.java
@@ -1,0 +1,11 @@
+package com.team.companyservice.company.application.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+@Getter
+public class UpdateCompanyActiveRequest {
+
+    @NotNull
+    private Boolean isActive;
+}
