+++ company-service/src/main/java/com/team/companyservice/company/domain/CompanyRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.team.companyservice.company.domain;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+import com.team.companyservice.company.application.search.CompanySearchCondition;
+
+public interface CompanyRepositoryCustom {
+    Page<Company> search(CompanySearchCondition condition, Pageable pageable);
+}
