+++ company-service/src/main/java/com/team/companyservice/company/domain/CompanyType.java
@@ -1,0 +1,6 @@
+package com.team.companyservice.company.domain;
+
+public enum CompanyType {
+    PRODUCER,
+    RECEIVER
+}
