--- company-service/src/main/java/com/team/companyservice/domain/company/CompanyRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.team.companyservice.domain.company;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import com.team.companyservice.application.company.CompanySearchCondition;
-
-public interface CompanyRepositoryCustom {
-    Page<Company> search(CompanySearchCondition condition, Pageable pageable);
-}
