--- company-service/src/main/java/com/team/companyservice/domain/company/CompanyType.java
@@ -1,6 +1,0 @@
-package com.team.companyservice.domain.company;
-
-public enum CompanyType {
-    PRODUCER,
-    RECEIVER
-}
