+++ company-service/src/main/java/com/team/companyservice/global/common/CurrentUser.java
@@ -1,0 +1,26 @@
+package com.team.companyservice.global.common;
+
+import java.util.UUID;
+
+public record CurrentUser(
+        UUID userId,
+        String role,
+        UUID hubId,
+        UUID companyId
+) {
+    public boolean isMasterAdmin() {
+        return "MASTER_ADMIN".equals(role);
+    }
+
+    public boolean isHubAdmin() {
+        return "HUB_ADMIN".equals(role);
+    }
+
+    public boolean isCompanyManager() {
+        return "COMPANY_MANAGER".equals(role);
+    }
+
+    public boolean isDeliveryManager() {
+        return "DELIVERY_MANAGER".equals(role);
+    }
+}
