+++ company-service/src/main/java/com/team/companyservice/global/config/FeignRetryConfig.java
@@ -1,0 +1,14 @@
+package com.team.companyservice.global.config;
+
+import feign.Retryer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class FeignRetryConfig {
+
+    @Bean
+    public Retryer feignRetryer() {
+        return new Retryer.Default(100, 1000, 3);
+    }
+}
