+++ company-service/src/main/java/com/team/companyservice/global/config/JpaAuditingConfig.java
@@ -1,0 +1,18 @@
+package com.team.companyservice.global.config;
+
+import java.util.Optional;
+import java.util.UUID;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+    @Bean
+    public AuditorAware<UUID> auditorProvider() {
+        return () -> Optional.of(UUID.fromString("00000000-0000-0000-0000-000000000001"));
+    }
+}
