+++ company-service/src/main/java/com/team/companyservice/global/error/ServiceException.java
@@ -1,0 +1,9 @@
+package com.team.companyservice.global.error;
+
+import com.team.common.exception.BusinessException;
+
+public class ServiceException extends BusinessException {
+    public ServiceException(CompanyErrorCode errorCode) {
+        super(errorCode);
+    }
+}
