--- company-service/src/main/java/com/team/companyservice/infrastructure/config/FeignRetryConfig.java
@@ -1,14 +1,0 @@
-package com.team.companyservice.infrastructure.config;
-
-import feign.Retryer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignRetryConfig {
-
-    @Bean
-    public Retryer feignRetryer() {
-        return new Retryer.Default(100, 1000, 3);
-    }
-}
