--- company-service/src/main/java/com/team/companyservice/presentation/common/CurrentUserResolverConfig.java
@@ -1,21 +1,0 @@
-package com.team.companyservice.presentation.common;
-
-import java.util.List;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class CurrentUserResolverConfig implements WebMvcConfigurer {
-
-    private final CurrentUserArgumentResolver currentUserArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(currentUserArgumentResolver);
-    }
-}
