--- company-service/src/main/java/com/team/companyservice/presentation/common/ServiceException.java
@@ -1,9 +1,0 @@
-package com.team.companyservice.presentation.common;
-
-import com.team.common.exception.BusinessException;
-
-public class ServiceException extends BusinessException {
-    public ServiceException(CompanyErrorCode errorCode) {
-        super(errorCode);
-    }
-}
