+++ company-service/src/test/java/com/team/companyservice/application/company/CompanyServiceTest.java
@@ -5,14 +5,17 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
-import com.team.companyservice.domain.company.Company;
-import com.team.companyservice.domain.company.CompanyRepository;
-import com.team.companyservice.domain.company.CompanyType;
+import com.team.companyservice.company.application.dto.request.CreateCompanyRequest;
+import com.team.companyservice.company.application.dto.response.CompanyResponse;
+import com.team.companyservice.company.application.service.CompanyService;
+import com.team.companyservice.company.domain.Company;
+import com.team.companyservice.company.domain.CompanyRepository;
+import com.team.companyservice.company.domain.CompanyType;
 import com.team.companyservice.infrastructure.client.HubClient;
-import com.team.companyservice.infrastructure.client.HubExistsResponse;
-import com.team.companyservice.presentation.common.CompanyErrorCode;
-import com.team.companyservice.presentation.common.CurrentUser;
-import com.team.companyservice.presentation.common.ServiceException;
+import com.team.companyservice.infrastructure.client.dto.HubExistsResponse;
+import com.team.companyservice.global.error.CompanyErrorCode;
+import com.team.companyservice.global.common.CurrentUser;
+import com.team.companyservice.global.error.ServiceException;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
